Text.Show.ByteString

Description

Efficiently convert from values to lazy byte strings.

The Show class

class Show a where

Conversion of values to readable byte strings. Minimal complete definition: showp or showpPrec

Minimal complete definition

Nothing

Methods

showpPrec :: Int -> a -> Put

Encodes a value to an efficient byte string builder. The precedence is used to determine where to put parentheses in a shown expression involving operators.

Values of type Put can be efficiently combined, so the showp functions are available for showing multiple values before producing an output byte string.

showp :: a -> Put

Encodes a value to an efficient byte string builder. Values of type Put can be efficiently combined, so this is available for building strings from multiple values.

showpList :: [a] -> Put

Allows for specialized display of lists of values. This is used, for example, when showing arrays of Chars.

show :: Show a => a -> ByteString

Encode a single value into a byte string

Putting Chars

putAscii :: Char -> Put

Writes a single Char to a byte string, assuming it's ascii.

putUTF8 :: Char -> Put

Writes a single Char to a byte string, properly UTF-8 encoded

Putting Strings

putAsciiStr :: String -> Put

Writes a string of ascii characters to a byte string

putUTF8Str :: String -> Put

Writes a string of unicode characters to a byte string, properly UTF-8 encoded

Putting digits

unsafePutDigit :: Int -> Put

Puts the decimal digit corresponding to the given Int without checking that it is in the interval [0,9]

putDigit :: Int -> Put

Puts the digit corresponding to the Int passed in.

Putting integers

showpIntAtBase :: Integral a => a -> (Int -> Char) -> a -> Put

Shows an Integral number using the base specified by the first argument and the chracter representation specified by the second.

Putting floats

showpGFloat :: RealFloat a => Maybe Int -> a -> Put

Show a signed RealFloat value using decimal notation when the absolute value lies between 0.1 and 9,999,999, and scientific notation otherwise. The optional integer can be used to specify precision.

showpFFloat :: RealFloat a => Maybe Int -> a -> Put

Show a signed RealFloat value using decimal notation. The optional integer can be used to specify precision.

showpEFloat :: RealFloat a => Maybe Int -> a -> Put

Show a signed RealFloat value using scientific (exponential) notation. The optional integer can be used to specify precision.

Combining builders

unlinesP :: [Put] -> Put

Merge several string builders, separating them by newlines

unwordsP :: [Put] -> Put

Merge several string builders, separating them by spaces

showpParen :: Bool -> Put -> Put

A utility function for surrounding output by parentheses conditionally.

Printing values

print :: Show a => a -> IO ()

Print a value to the standard output

Put

type Put = PutM ()

Put merely lifts Builder into a Writer monad, applied to ().

newtype PutM a :: * -> *

The PutM type. A Writer monad over the efficient Builder monoid.

Constructors

Put

Fields

unPut :: PairS a

runPut :: Put -> ByteString

Run the Put monad with a serialiser